requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
autopep8==2.0.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-debug-toolbar==4.1.0
django-ratelimit==4.0.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
fonttools==4.39.4
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
wsproto==1.2.0
zipp==3.15.0
gunicorn==20.1.0
waitress==2.1.2
django-cors-headers==3.10.0
PyJWT==2.8.0
xhtml2pdf==0.2.5
